@@ -4,32 +4,32 @@
   "private": true,
   "dependencies": {
     "@iconify/icons-octicon": "^1.2.5",
-    "@iconify/react": "^3.2.0",
-    "@reduxjs/toolkit": "^1.8.0",
-    "@testing-library/jest-dom": "^5.16.2",
-    "@testing-library/react": "^12.1.4",
-    "@testing-library/user-event": "^13.5.0",
-    "bootstrap": "5.1.3",
-    "chart.js": "^4.3.0",
-    "clsx": "^1.1.1",
-    "dateformat": "^5.0.3",
-    "deep-equal": "^2.0.5",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.2.1",
-    "react-chartjs-2": "^5.2.0",
-    "react-content-loader": "^6.2.0",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
-    "react-swipeable": "^6.2.1",
-    "redux": "^4.1.2",
+    "@iconify/react": "^4.0.0",
+    "@reduxjs/toolkit": "^1.9.5",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "bootstrap": "^5.3.2",
+    "chart.js": "^4.4.0",
+    "clsx": "^1.2.0",
+    "dateformat": "^5.0.4",
+    "deep-equal": "^2.1.1",
+    "react-bootstrap": "^2.8.0",
+    "react-chartjs-2": "^5.3.0",
+    "react-content-loader": "^6.5.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.1.0",
+    "react-router-dom": "^6.15.0",
+    "react-scripts": "5.0.1",
+    "react-swipeable": "^6.5.1",
+    "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
-    "web-vitals": "^2.1.4"
+    "web-vitals": "^3.1.0"
   },
   "scripts": {
     "start": "react-scripts start",
-    "build": "BUILD_PATH='./building' react-scripts build",
+    "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "version-major": "yarn version --major",
